fix random ci failures increasing timeout for start and log drain
