+++ openwhisk/initHandler_linux_test.go
@@ -1,5 +1,3 @@
-// +build linux
-
 /*
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
